Create a file for NodeNG see #587
